refactor: 사다리게임 입력 기능

+++ src/main/java/ladder/view/input/InputView.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface InputView {
     List<String> getUserNames();
-    Long getLadderHeight();
+    int getLadderHeight();
 }
